Move pool configs into pool package
